Changed constructor slightly

+++ src/metropolis.hpp
@@ -7,7 +7,7 @@
 
 class Metropolis {
 public:
-  Metropolis (int, int, int, double, double, double, std::string);
+  Metropolis (int, int, int, double, double, int, std::string);
   Metropolis (int, int, int, double, std::string);
   void run();
